\name{MR.Corr2-package}
\alias{MR.Corr2-package}
\alias{MR.Corr2}
\docType{package}
\title{\packageTitle{MR.Corr2}}
\description{\packageDescription{MR.Corr2}}
\details{
  The DESCRIPTION file: \packageDESCRIPTION{MR.Corr2}
  \packageIndices{MR.Corr2}

  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
  \packageAuthor{MR.Corr2}
  
  Maintainer: \packageMaintainer{MR.Corr2}
}
\references{
  This optional section can contain literature or other references for
  background information.
}
% Optionally other standard keywords, one per line,
% from the file KEYWORDS in the R documentation.
\keyword{package}
\seealso{
  Optional links to other man pages
}
\examples{
  ## Optional simple examples of the most important functions
  ## Use \dontrun{} around code to be shown but not executed
}
